pyproject.toml
[project]
name = "ezconstruct-api"
version = "0.1.0"
description = "FastAPI blueprint for large scale applications"
readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    { name = "Guilherme Rodrigues", email = "guilherme.b.rodrigues@skf.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "colorlog>=6.9.0",
    "fastapi[standard]>=0.115.13",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-json-logger>=3.3.0",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "httpx[http2]>=0.28.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
]
